# Conflicts:
#	keycloak-17.0.1/data/h2/keycloakdb.lock.db
#	keycloak-17.0.1/data/h2/keycloakdb.mv.db

+++ garage-microservice/src/main/java/com/tunidesign/garagemicroservice/service/GarageService.java
@@ -3,15 +3,14 @@ package com.tunidesign.garagemicroservice.service;
 
 import com.tunidesign.garagemicroservice.DTO.GarageRequestDTO;
 import com.tunidesign.garagemicroservice.DTO.GarageResponseDTO;
-import org.springframework.stereotype.Repository;
 
 import java.util.List;
 
 
 public interface GarageService {
     List<GarageResponseDTO> getGarages();
-    GarageResponseDTO getGarageById(int id);
+    GarageResponseDTO getGarageById(Long id);
     GarageResponseDTO save(GarageRequestDTO garageRequestDTO);
-    void deleteGarage(int id);
+    void deleteGarage(Long id);
     GarageResponseDTO update(GarageRequestDTO garageRequestDTO);
 }
