+++ ordre-microservice/src/main/java/com/tunidesign/ordremicroservice/service/OrderService.java
@@ -2,7 +2,6 @@ package com.tunidesign.ordremicroservice.service;
 
 import com.tunidesign.ordremicroservice.DTO.OrderRequestDTO;
 import com.tunidesign.ordremicroservice.DTO.OrderResponseDTO;
-import com.tunidesign.ordremicroservice.model.entity.Order;
 import org.springframework.web.multipart.MultipartFile;
 
 import java.io.IOException;
@@ -11,10 +10,10 @@ import java.util.List;
 public interface OrderService {
 
         List<OrderResponseDTO> getOrders();
-        OrderResponseDTO getOrderById(int id);
+        OrderResponseDTO getOrderById(Long id);
         OrderResponseDTO save(OrderRequestDTO orderRequestDTO);
-        void deleteOrder(int id);
+        void deleteOrder(Long id);
         OrderResponseDTO update(OrderRequestDTO orderRequestDTO);
 
-        OrderResponseDTO updatePhotos(MultipartFile img1 ,MultipartFile img2,MultipartFile img3,int id ) throws IOException;
+        OrderResponseDTO updatePhotos(MultipartFile img1 ,MultipartFile img2,MultipartFile img3,Long id ) throws IOException;
 }
